+++ PreguntIA/src/main/java/ObligatorioDDA_IS/Repository/SPGameRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ObligatorioDDA_IS.Models.SinglePlayerGame;
 
-public interface SPGameRepository extends JpaRepository<SinglePlayerGame, Integer> {
+public interface SPGameRepository extends JpaRepository<SinglePlayerGame, Long> {
     List<SinglePlayerGame> findByUserId(Long userId);
 }
